fix(Periphery): transfer excess from msg.sender to periphery and approve gClaimManager to withdraw it

+++ src/tests/test-helpers/User.sol
@@ -127,16 +127,16 @@ contract User {
         gClaimManager.exit(feed, maturity, balance);
     }
 
-    function doSwapTargetForZeros(address feed, uint256 maturity, uint256 balance, uint256 backfill, uint256 minAccepted) public {
+    function doSwapTargetForZeros(address feed, uint256 maturity, uint128 balance, uint256 backfill, uint256 minAccepted) public {
         periphery.swapTargetForZeros(feed, maturity, balance, backfill, minAccepted);
     }
-    function doSwapTargetForClaims(address feed, uint256 maturity, uint256 balance, uint256 minAccepted) public {
+    function doSwapTargetForClaims(address feed, uint256 maturity, uint128 balance, uint256 minAccepted) public {
         periphery.swapTargetForClaims(feed, maturity, balance, minAccepted);
     }
-    function doSwapZerosForTarget(address feed, uint256 maturity, uint256 balance, uint256 minAccepted) public {
+    function doSwapZerosForTarget(address feed, uint256 maturity, uint128 balance, uint256 minAccepted) public {
         periphery.swapZerosForTarget(feed, maturity, balance, minAccepted);
     }
-    function doSwapClaimsForTarget(address feed, uint256 maturity, uint256 balance, uint256 minAccepted) public {
+    function doSwapClaimsForTarget(address feed, uint256 maturity, uint128 balance, uint256 minAccepted) public {
         periphery.swapClaimsForTarget(feed, maturity, balance, minAccepted);
     }
